+++ src/weasyprint_libs/__init__.py
@@ -1,3 +1,0 @@
-from .loader import activate
-
-__all__ = ["activate"]